requirements-enhanced.txt
# Enhanced Backend API Requirements
# Next-generation dependencies for advanced caching, AI quality assurance, and performance optimization

# Core Flask dependencies (existing)
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Database dependencies (existing)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter for future migration

# Google APIs dependencies (existing)
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-cloud-language>=2.10.0
google-generativeai>=0.3.0

# Enhanced caching dependencies
redis>=4.5.0
redis-py-cluster>=2.1.0  # For Redis clustering support
hiredis>=2.2.0  # High-performance Redis parser

# AI and NLP enhancement dependencies
openai>=1.0.0  # For multi-model AI support
anthropic>=0.8.0  # Claude API integration
tiktoken>=0.5.0  # Token counting for AI models

# Performance and monitoring dependencies
prometheus-client>=0.17.0  # Metrics collection
structlog>=23.0.0  # Structured logging
psutil>=5.9.0  # System performance monitoring

# Data processing and validation
pydantic>=2.0.0  # Enhanced data validation
marshmallow>=3.20.0  # Schema serialization
jsonschema>=4.19.0  # JSON schema validation

# Async and concurrency support
asyncio-mqtt>=0.13.0  # Async support for future features
celery>=5.3.0  # Background task processing
kombu>=5.3.0  # Message queue support

# Security enhancements
cryptography>=41.0.0  # Enhanced encryption
pyotp>=2.9.0  # OTP generation for enhanced auth
argon2-cffi>=23.0.0  # Password hashing

# Testing and quality assurance
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Code linting
mypy>=1.5.0  # Type checking

# Development and debugging
flask-debugtoolbar>=0.13.0
memory-profiler>=0.61.0
py-spy>=0.3.0  # Production profiling

# Optional: Machine Learning for advanced features
scikit-learn>=1.3.0  # For content similarity analysis
transformers>=4.33.0  # Hugging Face transformers
sentence-transformers>=2.2.0  # Semantic similarity

# Data export and reporting
pandas>=2.0.0
openpyxl>=3.1.0  # Excel export
fpdf2>=2.7.0  # PDF generation
plotly>=5.15.0  # Interactive charts

# Rate limiting and API management
flask-limiter>=3.5.0
slowapi>=0.1.9  # Advanced rate limiting

# Configuration management
dynaconf>=3.2.0  # Advanced configuration management
click>=8.1.0  # CLI interface for management commands